fix: satisfy rebased lint checks
